refactor: Refactoring storybook of Modal Component
